refactor: move detectLocale to utilsFormat where its only usage lies
