+++ packages/ui/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './helpers'
 export * from './responsive'
 export * from './sharedStyles'
 export * from './utilsGas'
-export * from './utilsLocale'
 export * from './utilsFormat'
 export * from './utilsConnectState'
 export * from './utilsNetworks'
