+++ packages/ui/src/utils/utilsFormat.ts
@@ -1,11 +1,19 @@
+import { detect, fromUrl, fromNavigator } from '@lingui/detect-locale'
 import BigNumber from 'bignumber.js'
 import isUndefined from 'lodash/isUndefined'
 import isNaN from 'lodash/isNaN'
 
-import { detectLocale } from 'ui/src/utils/utilsLocale'
-
 BigNumber.config({ EXPONENTIAL_AT: 20, ROUNDING_MODE: BigNumber.ROUND_HALF_UP })
 export const BN = BigNumber
+
+export function detectLocale() {
+  if (typeof window !== 'undefined') {
+    const DEFAULT_FALLBACK = () => 'en'
+    return detect(fromUrl('lang'), fromNavigator(), DEFAULT_FALLBACK) ?? 'en-US'
+  } else {
+    return 'en-US'
+  }
+}
 
 const localeDetected = detectLocale() || 'en-US'
